job/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.3
gevent
pydantic==2.5.2
python-dotenv
python-jose
python-multipart==0.0.9
google-cloud-pubsub
SQLAlchemy
uvicorn
pg8000